requirements.txt
flask>=2.0.1
python-dotenv>=0.19.0
pandas>=1.3.0
plotly>=5.3.1
requests>=2.31.0
websocket-client>=1.6.1
gunicorn>=20.1.0
Flask-SocketIO
eventlet
kiteconnect>=3.9.2
